Do not use HTTP/3 for now
There seems to be a bug in Caddy.
When a new config is set because an app is deployed,
if you're in Chrome and keep reloading, you'll keep getting
responses from the old Caddy server. And since the old Caddy
server points to the previous deployment, there app is not
running anymore, so the old Caddy returns 502.

In other words, the connection is kept open even if there's
currently no un-responded requests. Waiting ~15-30 seconds
without reloading would let the old server stop and
'fix' the issue.