Add extraSwarmParams
The goal is to have an escape hatch when users have
very specific needs that Disco does not currently
support and may or may not eventually support.

It does expose the API of Docker Swarm
and should be considered somewhat unstable,
in the sense that we don't control the API
of Docker Swarm, and also that Disco may
eventually use something else than Docker Swarm.